feat: 게시글 조회수 기능 구현

+++ src/main/java/codesquad/springcafe/db/ArticleDatabase.java
@@ -14,14 +14,14 @@ import java.util.Map;
 public class ArticleDatabase {
     private static final Logger logger = LoggerFactory.getLogger(ArticleDatabase.class);
 
-    private Map<Integer, Article> articles = new LinkedHashMap<>();
+    private Map<Long, Article> articles = new LinkedHashMap<>();
 
-    public int getArticleID() {   // Article에 ID를 주기 위한 메소드
+    public long getArticleID() {   // Article에 ID를 주기 위한 메소드
         return articles.size() + 1;    // index는 1번부터 시작하도록 +1
     }
 
     public void addArticle(Article article) {
-        int articleId = getArticleID();
+        long articleId = getArticleID();
         article.setArticleId(articleId);    // memory에서는 직접 setter를 통해 할당
         articles.put(articleId, article);
         logger.debug("Article # {} ,[{}] Added at Memory DB", articleId, article);
@@ -31,7 +31,7 @@ public class ArticleDatabase {
         return new ArrayList<>(articles.values());
     }
 
-    public Article findArticleById(int articleId) {
+    public Article findArticleById(long articleId) {
         logger.debug("articleId  :{}", articleId);
         if (articleId >= 0 && articleId <= articles.size()) {
             return articles.get(articleId);
